Give the 28-bit address and the hex formatters an explicit width
`long` is 32 bits on the machine and 64 on a host, so every host test that
touched an Addr28 was working with a wider number than the target has: an
address that overflows 28 or 32 bits there quietly does not, and the check that
would have caught it passes. hex_to_buf has the same gap the other way, shifting
right by 28 to reach a top nibble that a 64-bit value does not keep there.

int32_t is the same type as long on this target, so nothing about the generated
code changes; it is only on the host that it now says what the machine means.
The static_assert states the property where a future edit would trip over it.

Left alone: the lcopy casts. Those match mega65-libc's own signatures, and
uint32_t there was measured at 694 bytes for a sign bit the addresses never use.

  .M65  no change

+++ src/format.c
@@ -16,7 +16,7 @@ uint8_t nybl_to_screen(uint8_t v) {
 }
 
 /* Always eight digits; format_hex() takes the tail it wants. */
-static void hex_to_buf(char* out, const long value) {
+static void hex_to_buf(char* out, const int32_t value) {
     out[0] = nybl_to_screen((uint8_t)(value >> 28));
     out[1] = nybl_to_screen((uint8_t)(value >> 24));
     out[2] = nybl_to_screen((uint8_t)(value >> 20));
@@ -29,7 +29,7 @@ static void hex_to_buf(char* out, const long value) {
 
 /* Writes the low `columns` digits.  Takes a pointer rather than an address as
  * an integer, which would hide the stores from the compiler. */
-void format_hex(char* out, const long value, const char columns) {
+void format_hex(char* out, const int32_t value, const char columns) {
     char dec[8];
     hex_to_buf(dec, value);
 
@@ -74,7 +74,7 @@ char* append_str(char* at, const char* text) {
 
 /* format_hex() writes `columns` digits and no terminator, which is what lets
  * these chain from a cursor. */
-char* append_hex(char* at, long value, uint8_t columns) {
+char* append_hex(char* at, int32_t value, uint8_t columns) {
     format_hex(at, value, (char)columns);
     return at + columns;
 }
